+++ include/H2DE/surfaces/H2DE_texture.h
@@ -24,7 +24,7 @@ private:
     H2DE_Texture(H2DE_Engine* engine, const H2DE_SurfaceData& sd, const H2DE_TextureData& ted);
     ~H2DE_Texture() override;
 
-    std::optional<H2DE_AbsRect> getSrcRect() const override;
+    inline std::optional<H2DE_AbsRect> getSrcRect() const override { return ted.srcRect; };
     
 public:
     /**